Fix warning on Windows build
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk/activemq-cpp@925699 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/activemq/wireformat/openwire/utils/BooleanStream.cpp
@@ -106,7 +106,7 @@ void BooleanStream::marshal( DataOutputStream* dataOut ) throw ( IOException ) {
         }
 
         // Dump the payload
-        dataOut->write( &data[0], data.size(), 0, arrayLimit );
+        dataOut->write( &data[0], (int)data.size(), 0, arrayLimit );
         clear();
     }
     AMQ_CATCH_RETHROW( IOException )
